folder refactor: tests/ now match up.
