+++ package.json
@@ -10,6 +10,7 @@
     "closure": "closure/closure-type-checking.js",
     "lint": "eslint .",
     "test": "npm run lint --silent && npm run unit && npm run closure",
+    "watch": "find . -name '*.js' | grep -v 'node_modules' | grep -v 'extension' | entr npm run unit",
     "unit": "mocha $(find ./test -name '*.js') --timeout 60000",
     "chrome": "./scripts/launch-chrome.sh",
     "smoke": "./scripts/run-smoke-tests.sh",
