+++ test/fixtures/manifest.json
