+++ test/src/audits/manifest/icons.js
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-const Audit144 = require('../../../src/audits/manifest/icons-min-144.js');
-const Audit192 = require('../../../src/audits/manifest/icons-min-192.js');
+const Audit144 = require('../../../../src/audits/manifest/icons-min-144.js');
+const Audit192 = require('../../../../src/audits/manifest/icons-min-192.js');
 const assert = require('assert');
-const manifestParser = require('../../../src/lib/manifest-parser');
+const manifestParser = require('../../../../src/lib/manifest-parser');
 
 /* global describe, it*/
 
@@ -58,7 +58,7 @@ describe('Manifest: icons audits', () => {
 
     it('succeeds when a manifest contains icons that are large enough', () => {
       // stub manifest contains a 192 icon
-      const manifestSrc = JSON.stringify(require('./manifest.json'));
+      const manifestSrc = JSON.stringify(require('../../../fixtures/manifest.json'));
       const manifest = manifestParser(manifestSrc);
       assert.equal(Audit144.audit({manifest}).value, true);
       assert.equal(Audit192.audit({manifest}).value, true);
